refactor: split SettingsPage.tsx into per-tab containers (closes #493)
